Fix entire db flush when no "only" keys found
Resolves bug that would wipe the entire Redis db when an `only`
specification is provided but there are no matching keys in the db.
This is achieved by defaulting the `only` and `except` arguments to
be `nil`, which signals that they have not been set whereas
implementation before was ambiguous. And if `only` has not been set,
then the code knows to use all Redis keys.